Fix float printing discrepancy. (#23)
* Replaces float with double everywhere.
* Changes every runtime conversion from float to string to round to one decimal place with ".1f" (this will probably change later. I just need the acceptance tests to pass and I don't really care about float precision.)
* The auto-formatter moved a bunch of stuff around on save.

+++ src/codegen.cpp
@@ -349,7 +349,7 @@ CompNodeResult gen_int_literal(ASTNode &node, CompSymbolTable &st) {
 }
 
 CompNodeResult gen_float_literal(ASTNode &node, CompSymbolTable &st) {
-  auto value = node.data.at("value").get<float>();
+  auto value = node.data.at("value").get<double>();
   std::stringstream ss;
   ss << "make_float(" << value << ")";
   auto s = ss.str();
